Fixed JSON validation dependency.
It appears that it was actually correct in the course handout pdf. I
copy/pasted it wrong by using another dependency as a template for
copying and then accidentally left in a <>test<> element. That seemed to
be what was breaking it and preventing Spring from finding/using the
validation API.