Added /update functionality.
JSON must be passed as { "id":5,"firstName":"Gustavo"}. Not sure why the
"firstName" version is chosen or from where (service or student).
Also, id has NotNull validation which gives errors, plural, in the video
along with the error message we set but when I run it in postman all I
get is a simple 400 Bad Request (without all the errors and the custom
error message).